+++ data/src/main/java/eu/batomobile/data/model/dto/VideoDTO.kt
@@ -5,8 +5,8 @@ import com.google.gson.annotations.SerializedName
 
 internal data class VideoDTO(
     @SerializedName("id") var id: String? = null,
-    @SerializedName("startTime") var startTime: Int? = null,
-    @SerializedName("endTime") var endTime: Int? = null,
+    @SerializedName("startTime") var startTime: Long? = null,
+    @SerializedName("endTime") var endTime: Long? = null,
     @SerializedName("service") var service: String? = null,
     @SerializedName("fullUrl") var fullUrl: String? = null
 )
